First revision from user feedback
Bug when try to save any date, fixed.
Add new field, status_produksi, checkbox field.
Bug in routing url, fixed

+++ controller/Status_Spk.php
@@ -35,10 +35,13 @@ class Status_Spk extends Controller {
 			$data = $this->_mSpk->filterByDate($df, $dt);
 		} else $data = $this->_searchSpk($sf, $kw);
 		
-		$heading = $this->_getHeading(7);
+		$heading = $this->_getHeading(8);
 		$heading[] = 'action';
 		
 		$fullHeading = $this->_getHeading(99);
+		
+		$data = $this->_reformatDates($data);
+		$data = $this->_proccessStatusProduksi($data);
 		
 		$viewData = array(
 			'full_heading' => $fullHeading,
@@ -53,7 +56,8 @@ class Status_Spk extends Controller {
 	{
 		foreach ($data as $row => $value)
 		{
-			$id = $value['nomor'];
+			$id      = $value['nomor'];
+			$tanggal = $value['tanggal'];
 			
 			$editLink   = "http://" . Base::site_url('/spk?id=' . $id);
 			$chooseLink = "http://" . Base::site_url('/cetak?id=' . $id);
@@ -64,8 +68,8 @@ class Status_Spk extends Controller {
 			$action = <<< PHREDOC
 <div class="btn-group" style="width: 107px;">
 	<a class="btn" href="$editLink"><i class="icon-pencil"></i></a>
-	<a class="btn" href="$chooseLink"><i class="icon-active"></i></a>
-	<a class="btn" href="$deleteLink"><i class="icon-trash"></i></a>
+	<a class="btn" href="$chooseLink"><i class="icon-printer"></i></a>
+	<a class="btn" href="$deleteLink" onclick="return confirm('Apakah Anda yakin ingin menghapus SPK No. $id tanggal $tanggal ini?');"><i class="icon-trash"></i></a>
 </div>
 PHREDOC;
 			$data[$row]['action'] = $action;
@@ -110,11 +114,83 @@ PHREDOC;
 		$status = Base::savecho($g['status'], '', true);
 		
 		if(!empty($id)) {
+			
 			$this->_mSpk->changeStatus($id, $status);
+			
 			echo json_encode(array('result' => true));
+		
 		} else {
-			echo json_encode(array('result' => false));
+			
+			$spdata = Base::savecho($g['spdata'], '', true);
+			
+			if(!empty($spdata)) {
+				
+				$spData = json_decode($spdata);
+				
+				$id = $spData->id;
+				
+				$statusProduksi = $spData->sp;
+				
+				$this->_mSpk->changeStatusProduksi($id, $statusProduksi);
+				
+				echo json_encode(array('result' => true));
+				
+			} else 
+				echo json_encode(array('result' => false));
 		}
+	}
+	
+	private function _reformatDates($arrData)
+	{
+		foreach ($arrData as $key => $data) {
+			$arrData[$key] = Base::reformat_date($data, "d-m-Y", $this->_mSpk->dateFields);
+		}
+		
+		return $arrData;
+	}
+	
+	private function _proccessStatusProduksi($data)
+	{
+		foreach ($data as $key => $row) {
+			
+			$sp = isset($row['status_produksi']) && !empty($row['status_produksi']) ? 
+				$row['status_produksi'] : "[]";
+			
+			$arrSp = json_decode($sp, true);
+			
+			if($arrSp == null) $arrSp = array();
+			
+			$last = end($arrSp);
+			
+			$nomor = $row['nomor'];
+			
+			$data[$key]['status_produksi'] = $this->_createSpUserControl($last, $sp, $nomor);
+		}
+		
+		return $data;
+	}
+	
+	private function _createSpUserControl($caption, $jsonStatus, $id)
+	{
+		$caption = empty($caption) ? 'New' : $caption;
+		
+		$html = '<a id="sp_caption_' . $id . '" href="#sp_caption_' . $id . '" onclick="app.showStatusProduksiCheck(' . "'$id'" . ')">' . $caption . '</a>';
+		
+		$cbx = $this->_mSpk->createStatusProduksiCheckbox($jsonStatus);
+		
+		$html .= <<< PHREDOC
+<div id="sp_control_$id" style="display: none; margin-top: 5px;">
+	<div>$cbx</div>
+	<div style="margin-top: 5px;">
+		<div class="btn-group">
+			<a style="width: 30px;" class="btn btn-info" onclick="app.closeStatusProduksiCheck('$id', true)">OK</a>
+			<a style="width: 30px;" class="btn btn-warning" onclick="app.closeStatusProduksiCheck('$id')">Batal</a>
+		</div>
+	</div>
+<div>
+PHREDOC;
+		
+		return $html;
 	}
 }
